Change SqmPrice type from decimal to int
Updated the calculation of the average square meter price
to use an integer data type instead of decimal. This change
may impact the precision of the average calculation.

+++ BusinessLogic/AssessmentService.cs
@@ -16,7 +16,7 @@ namespace BusinessLogic
             {
                 var enumDataTable = dataTable.AsEnumerable();
 
-                var averageSqmPrice = enumDataTable.Select(row => row.Field<decimal>("SqmPrice")).Average();
+                var averageSqmPrice = enumDataTable.Select(row => row.Field<int>("SqmPrice")).Average();
 
                 return Convert.ToInt32(averageSqmPrice) * priceAssessment.SquareMeters;
             }
